Add Pluto Deprecated fields
These fields DO exist on the OpenData page, but don't exist in the
dataset itself. Rather than change the schema on OpenData, I'm just
punting to get PLUTO distribute'able.

Also changing the type of the test PLUTO page so that it doesn't dist
with the rest. We might just want to delete this page.